SeqLogo

Documentation for SeqLogo.

SeqLogo.save_crosslinked_logoplot

SeqLogo.save_logoplot

SeqLogo.save_logoplot

SeqLogo.save_crosslinked_logoplot — Method

save_crosslinked_logoplot(pfm, background, c, save_name; dpi=65, rna=true)
Save a logoplot with crosslinks to a file

SeqLogo.save_logoplot — Method

save_logoplot(pfm, background, save_name; dpi=65)

Arguments

pfm::Matrix{Real}: Position frequency matrix

background::Vector{Real}: Background probabilities of A, C, G, T

save_name::String: Name of the path/file to save the plot

Note that

pfm must be a probability matrix

sum of each column must be 1

background must be a vector of length 4

must be a vector of probabilities

sum of background must be 1

Example

pfm =  [0.01  1.0  0.98  0.0   0.0   0.0   0.98  0.0   0.18  1.0
        0.98  0.0  0.01  0.19  0.0   0.96  0.01  0.89  0.03  0.0
        0.0   0.0  0.0   0.77  0.01  0.0   0.0   0.0   0.56  0.0
        0.0   0.0  0.0   0.05  0.99  0.04  0.01  0.11  0.24  0.0]

background = [0.25, 0.25, 0.25, 0.25]

#= save the logo plot in the tmp folder as logo.png =#
save_logoplot(pfm, background, "tmp/logo.png")

#= save the logo plot in the current folder as logo.png with a dpi of 65 =#
save_logoplot(pfm, background, "logo.png"; dpi=65)

SeqLogo.save_logoplot — Method

save_logoplot(pfm, save_name; dpi=65)

This is the same as `save_logoplot(pfm, background, save_name; dpi=65)`
where `background` is set to `[0.25, 0.25, 0.25, 0.25]`

See `save_logoplot(pfm, background, save_name; dpi=65)` for more details.